v2 - sumission approve word complete

+++ database/migrations/2024_06_19_072030_add_new_columns_to_journal_board_members_table.php
@@ -1,22 +1,26 @@
 <?php
+
 use Illuminate\Database\Migrations\Migration;
 use Illuminate\Database\Schema\Blueprint;
 use Illuminate\Support\Facades\Schema;
 
 class AddNewColumnsToJournalBoardMembersTable extends Migration
 {
-    /**
-     * Run the migrations.
-     *
-     * @return void
-     */
     public function up()
     {
         Schema::table('journal_board_members', function (Blueprint $table) {
-            $table->string('google_scholar')->nullable()->after('journal_id');
-            $table->string('scopus')->nullable()->after('google_scholar');
-            $table->string('orcid')->nullable()->after('scopus');
-            $table->string('email')->nullable()->after('name');
+            if (!Schema::hasColumn('journal_board_members', 'google_scholar')) {
+                $table->string('google_scholar')->nullable()->after('journal_id');
+            }
+            if (!Schema::hasColumn('journal_board_members', 'scopus')) {
+                $table->string('scopus')->nullable()->after('google_scholar');
+            }
+            if (!Schema::hasColumn('journal_board_members', 'orcid')) {
+                $table->string('orcid')->nullable()->after('scopus');
+            }
+            if (!Schema::hasColumn('journal_board_members', 'email')) {
+                $table->string('email')->nullable()->after('name');
+            }
             // Add more columns as needed
         });
     }
@@ -29,10 +33,18 @@ class AddNewColumnsToJournalBoardMembersTable extends Migration
     public function down()
     {
         Schema::table('journal_board_members', function (Blueprint $table) {
-            $table->dropColumn('google_scholar');
-            $table->dropColumn('scopus');
-            $table->dropColumn('orcid');
-            $table->dropColumn('email');
+            if (Schema::hasColumn('journal_board_members', 'google_scholar')) {
+                $table->dropColumn('google_scholar');
+            }
+            if (Schema::hasColumn('journal_board_members', 'scopus')) {
+                $table->dropColumn('scopus');
+            }
+            if (Schema::hasColumn('journal_board_members', 'orcid')) {
+                $table->dropColumn('orcid');
+            }
+            if (Schema::hasColumn('journal_board_members', 'email')) {
+                $table->dropColumn('email');
+            }
             // Drop additional columns if needed
         });
     }
